refactor: split ScrapeExecutor behind facade
ScrapeExecutor.ts was 167 effective LoC, over the Phase 12e
max-lines:150 cap. Split the implementation into four cohesive
sub-modules under Executor/ (Types, Fetch, Account, Execute) and
turn ScrapeExecutor.ts into a barrel facade that re-exports the
unchanged public surface (executeScrape, default + named).

Pure behavior-preserving code-motion: the three ScrapeExecutor
unit suites (21 tests) pass unmodified. The max-lines:150 cap for
Strategy/Scrape lands in a separate chore(eslint) commit.
